[anhalize_*] Moving class for future developement.
