+++ anhalyze/core/anhalyze_utils.py
@@ -14,7 +14,6 @@ import deprecation
 
 # Project custom made libraries
 import anhalyze.core.anhalyze_plot_utils as apu
-from .anhalyze import get_date
 from anhalyze.core.anhalyze_geo import getIndex_sec
 
 
@@ -79,143 +78,6 @@ def get_paths(run_name=None, environ_paths=False):
 
 # TODO need to move bunch of functions out of here
 #   also need to separate io, timeseries, and whatever else needed
-
-
-@deprecation.deprecated()
-class Anhalyze:
-    """ This class will do analysis of ANHA4 data, for now it initializes the location of files.
-   ...
-
-    Attributes
-    ----------
-        run_name : str
-            Simulation name (e.g. 'ANHA4-WJM004')  (default is ??)
-        grid : str
-            Simulation output grid. (e.g. gridT, gridU)   (default is T)
-        years : str
-            List of years to be analysed  (default is [1998])
-        month_list : str
-            List of months to be analysed. If None(default), return all 12 months
-        one_per_month : bool
-            If True return only the first file of each month (default is False)
-        verbose : bool
-            Print outputs (default is True)
-
-    Methods
-    -------
-    get_file_list(month_list=None, one_per_month=False, monthly_mean=False)
-        Returns file list given a list of years, a grid type,
-        and either all the days in a month, or the first one.
-
-    """
-
-    def __init__(self, run_name=None, grid=None, years=None, month_list=None, one_per_month=False, verbose=True):
-        """ Initializing class.
-
-        Parameters
-        ----------
-        run_name : str, optional
-            Simulation name (e.g. 'ANHA4-WJM004')  (default is ??)
-        grid : str, optional
-            Simulation output grid. (e.g. gridT, gridU)   (default is T)
-        years : str, optional
-            List of years to be analysed  (default is [1998])
-        month_list : str, optional
-            List of months to be analysed. If None(default), return all 12 months
-        one_per_month : bool, optional
-            If True return only the first file of each month (default is False)
-        verbose : bool, optional
-            Print outputs (default is True)
-
-         TODO: where is the nc.Dataset to get the data?
-        """
-
-        # -------
-        # The global wild west
-        self.run_name = run_name
-        if grid is None:
-            self.grid = 'T'
-        else:
-            self.grid = grid
-        # TODO: add assert to grid values
-        if years is None:
-            self.years = ['1998']
-        else:
-            self.years = years
-        # TODO: add assert to years
-        self.month_list = month_list
-        self.one_per_month = one_per_month
-        self.verbose = verbose
-
-        # Init file list given conditions
-        self.file_list = self.get_file_list(one_per_month=one_per_month, month_list=month_list)
-
-        if verbose:
-            print(self.file_list)
-
-    def get_file_list(self, month_list=None, one_per_month=False, monthly_mean=False):
-        """  Returns file list given a list of years, a grid type,
-             and either all the days in a month, or the first one.
-
-        Parameters
-        ----------
-        month_list :  str, optional
-            List of months to be analysed. If None(default), return all 12 months
-        one_per_month :  bool, optional
-            If True return only the first file of each month (default is False)
-        monthly_mean :  bool, optional
-            If not using the default 5 days averaged outputs,
-            The function ask the user to insert the monthly mean path.
-
-        Returns
-        -------
-        selected_file_list: list
-            List of files for analysis
-
-        """
-
-        # Setup
-        selected_file_list = []
-        monthly_file_list = []
-        if month_list is None:
-            month_list = []
-
-        # Get paths
-        data_path, mask_path = get_paths(self.run_name)
-
-        # Get complete file list from path
-        file_list = os.listdir(data_path)
-
-        # Selecting list of files given params
-        for year in self.years:
-            selected_file_list += (sorted([f for f in file_list if '_y' + year in f and '_grid' + self.grid in f]))
-
-        # Selecting first day on given month
-        if one_per_month:
-            if not month_list:
-                month_list = [get_date(filename, how='m') for filename in selected_file_list]
-
-            for year in self.years:
-                for month in month_list:
-                    file_name_stump = 'y{}m{}'.format(year, month)
-                    file_month_name = [f for f in selected_file_list if file_name_stump in f][0]
-                    monthly_file_list.append(file_month_name)
-        else:
-            # Make month selection
-            if month_list:
-                for year in self.years:
-                    for month in month_list:
-                        file_name_stump = 'y{}m{}'.format(year, month)
-                        file_month_names = [f for f in selected_file_list if file_name_stump in f]
-                        monthly_file_list += file_month_names
-
-        if monthly_file_list:
-            selected_file_list = monthly_file_list
-
-        # Adding full path to filenames
-        selected_file_list = [data_path + filename for filename in selected_file_list]
-
-        return selected_file_list
 
 
 @deprecation.deprecated()
